Further split off tables factory for building outputs as per #316
This work also lays the groundwork for GHIs #240 #241 & #242

Just the one function initially, later we can put table utility classes here too.
